h2_client: narrow continuation_count to u8

+++ src/http/h2_client/ClientSession.rs
@@ -73,7 +73,7 @@ pub struct ClientSession {
     /// Stream id whose CONTINUATION sequence is in progress; 0 = none.
     pub expecting_continuation: u31,
     /// CONTINUATION frames seen so far in the current header block.
-    pub continuation_count: u32,
+    pub continuation_count: u8,
 
     /// Cold-start coalesced requests parked until the server's first SETTINGS
     /// frame arrives so the real MAX_CONCURRENT_STREAMS cap can be honoured.
